+++ demo/src/assets/IMaterial.h
@@ -1,0 +1,15 @@
+#pragma once
+
+namespace Assets
+{
+	class IMaterial
+	{
+	public:
+		virtual ~IMaterial() = default;
+		class IInstance
+		{
+		public:
+			virtual ~IInstance() = default;
+		};
+	};
+}
